% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juicedown-package.R
\docType{package}
\name{juicedown-package}
\alias{juicedown}
\alias{juicedown-package}
\title{juicedown: juice+markdown to Generate HTML with Inline Styles from R Markdown v1}
\description{
A convenient tool to create HTML with inline styles. It is particularly useful when working on CMS whose code editor eliminates style and link tags. The main use case of the package is the learning management system, Moodle. Additional helper functions for teaching purposes are provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kenjisato.github.io/juicedown/}
}

}
\author{
\strong{Maintainer}: Kenji Sato \email{kenji@kenjisato.jp} (\href{https://orcid.org/0009-0000-4520-2560}{ORCID})

}
\keyword{internal}
