% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juicydown-package.R
\docType{package}
\name{juicydown-package}
\alias{juicydown}
\alias{juicydown-package}
\title{juicydown: Convert R Markdown into HTML with Inline Styles}
\description{
A convenient tool to create HTML with inline styles. It is particularly useful when working on CMS whose code editor eliminates style and link tags. The main use case of the package is the learning management system, Moodle.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kenjisato.github.io/juicydown/}
}

}
\author{
\strong{Maintainer}: Kenji Sato \email{kenji@kenjisato.jp} (\href{https://orcid.org/0009-0000-4520-2560}{ORCID})

}
\keyword{internal}
