% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweaks.R
\name{tweak_footnote_highlight}
\alias{tweak_footnote_highlight}
\title{Highlight target link for footnotes}
\usage{
tweak_footnote_highlight()
}
\value{
Script block that makes highlighting possible.
}
\description{
Highlight target link for footnotes
}
