% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweaks.R
\name{tweak_moodle_heading}
\alias{tweak_moodle_heading}
\title{Hide H2 heading for Moodle LMS}
\usage{
tweak_moodle_heading()
}
\value{
Script block that makes H2 head of the content invisible.
}
\description{
You may want the title of the article to be different from the title of the
resource. Call this function in a code chunk somewhere in your R Markdown file.
}
