% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juicedown-package.R
\docType{package}
\name{juicedown-package}
\alias{juicedown}
\alias{juicedown-package}
\title{juicedown: 'juice' + 'markdown': Convert 'R Markdown' into 'HTML' with Inline Styles}
\description{
A convenience tool to create 'HTML' with inline styles using 'juicyjuice' and 'markdown' packages. It is particularly useful when working on a content management system (CMS) whose code editor eliminates style and link tags. The main use case of the package is the learning management system, 'Moodle'. Additional helper functions for teaching purposes are provided. Learn more about 'juicedown' at \url{https://kenjisato.github.io/juicedown/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kenjisato.github.io/juicedown/}
}

}
\author{
\strong{Maintainer}: Kenji Sato \email{kenji@kenjisato.jp} (\href{https://orcid.org/0009-0000-4520-2560}{ORCID}) [copyright holder]

}
\keyword{internal}
